Create 4-way sort
This code defines a SortingGameViewController class that displays an image and listens for swipe gestures from the user. If the swipe direction matches the correct answer for the current image, the score is updated and a new image is displayed. If the swipe direction is incorrect, an alert is shown to inform the user.