Create iOS sign in with ethereum
Sample code in Swift that you can use to implement a sign up flow for an iOS app that allows users to create an account using their Ethereum address.

This code uses the EthereumKit framework to generate a new Ethereum address and a private recovery key for the address. It then displays the private recovery key to the user and prompts them to write it down on a piece of paper. A button is provided for the user to confirm that they have written down their private recovery key. When the button is tapped, the app can proceed to the next screen.